Consolidating log functions from app to RunInstanceLogger
