+++ test/cli/fixtures/empty-run-logs.json
@@ -1,0 +1,1 @@
+[]
