+++ test/cli/fixtures/run-logs.json
@@ -1,0 +1,1 @@
+[{"dateRun":"2021-06-17T15:10:49.601Z","toDate":"2020-06-16","fromDate":"2019-01-01"}]
